src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
slack-bolt
python-telegram-bot[ext]
apscheduler
google-generativeai
requests